Move the reference stack into its own file
